feat(App.tsx, index.tsx): update app and index to tsx and update code to work with react 18.x
