+++ src/App.tsx
@@ -1,13 +1,8 @@
-import React from 'react';
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route
-} from "react-router-dom";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
-import './App.css';
-import Home from './pages/homepage/HomePage';
-import About from './pages/aboutpage/AboutPage';
+import "./App.css";
+import Home from "./pages/homepage/HomePage";
+import About from "./pages/aboutpage/AboutPage";
 import Footer from "./components/Footer.component";
 import { ToastContainer } from "react-toastify";
