--- src/index.js
@@ -1,28 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { createTheme, ThemeProvider, StyledEngineProvider } from "@mui/material/styles";
-import "./index.css";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
-
-const theme = createTheme({
-  typography: {
-    fontFamily: "'Playfair Display', serif"
-  }
-});
-
-ReactDOM.render(
-  <React.StrictMode>
-    <StyledEngineProvider injectFirst>
-      <ThemeProvider theme={theme}>
-        <App />
-      </ThemeProvider>
-    </StyledEngineProvider>
-  </React.StrictMode>,
-  document.getElementById("root")
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
